pyproject.toml
[project]
name = "ocorrencias"
version = "0.1.0"
description = "Classificação de ocorrências aeronáuticas"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "guardrails-ai>=0.5.15",
    "openai>=1.57.0",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
]
